Split code into equation editor and functional derived component editor
